Аномально большое время копирования в тесте simple_test()

+++ copy_between_streams_with_speed/src/androidTest/java/com/github/aakumykov/copy_between_streams_with_speed/TestBase.kt
@@ -106,7 +106,10 @@ abstract class TestBase {
             writeAndDisplay(random.nextBytes(additionalBytesCount))
         }
 
-        Assert.assertEquals(dataSizeBytes, file.length())
+        Assert.assertEquals(
+            dataSizeBytes.toLong(),
+            file.length()
+        )
 
         println("writeTestDataToFile(${dataSizeBytes.humanDecimalPlaces}) ФИНИШ")
     }
